requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
google-genai
python-dotenv